+++ src/utils.js
@@ -24,9 +24,14 @@ function isRegExp(value) {
     return toString.call(value) === '[object RegExp]';
 }
 
+function isArrayLike(value) {
+    return value && hasOwnProperty.call(value, 'length');
+}
+
 module.exports = {
     addToSet,
     getPropertyValue,
     isPlainObject,
-    isRegExp
+    isRegExp,
+    isArrayLike
 };